revisi
- cek foreign key ke artikel pas delete
- cek ada ato ga pas get by kode kategori
- cek nama apa ada yang sama pas insert kategori
- cek nama apa ada yang sama pas update kategori

+++ routes/admin/kategori.js
@@ -8,6 +8,10 @@ router.get('/', async (req, res) => {
 });
 
 router.get('/:kategori_id', async (req, res) => {
+    if(! await kategori_model.ifExists("kategoris", "kategori_id", kategori_id)){
+        return res.status(404).send({error: "Kategori not found"});
+    }
+
     const result = await kategori_model.getDataById(req.params.kategori_id);
     return res.status(200).send(result[0]);
 });
@@ -17,6 +21,10 @@ router.post('/', async (req, res) => {
     const data = {
         kategori_nama : nama
     };
+
+    if(await kategori_model.ifExists("kategoris", "kategori_nama", nama)){
+        return res.status(500).send({error: "Duplicate names"});
+    }
 
     if(await kategori_model.insertData(data)){
         return res.status(201).send(data);
@@ -32,6 +40,10 @@ router.put('/:kategori_id', async (req, res) => {
         return res.status(404).send({error: "Kategori not found"});
     }
 
+    if(await kategori_model.ifExists("kategoris", "kategori_nama", nama)){
+        return res.status(500).send({error: "Duplicate names"});
+    }
+
     const result = await kategori_model.updateData(kategori_id, nama);
     return res.status(200).send(result);
 });
@@ -40,6 +52,10 @@ router.delete('/:kategori_id', async (req, res) => {
     const { kategori_id } = req.params;
     if(! await kategori_model.ifExists("kategoris", "kategori_id", kategori_id)){
         return res.status(404).send({error: "Kategori not found"});
+    }
+
+    if( await kategori_model.ifExists("artikels", "kategori_id", kategori_id)){
+        return res.status(500).send({error: "Kategori used by some articles"});
     }
 
     if(await kategori_model.deleteData(kategori_id)){
